+++ app/layout.tsx
@@ -1,0 +1,10 @@
+import './globals.css'
+import type { ReactNode } from 'react'
+
+export default function RootLayout({ children }: { children: ReactNode }) {
+  return (
+    <html lang="en">
+      <body className="min-h-screen bg-background">{children}</body>
+    </html>
+  )
+}
